Fix validation of transaction value

+++ src/main/java/com/rinha/backend/entity/Transaction.java
@@ -11,7 +11,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigInteger;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -24,7 +24,7 @@ public class Transaction {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
     private String type;
-    private BigInteger value;
+    private BigDecimal value;
     private String description;
     @ManyToOne
     @JoinColumn(name = "client_id")
